ONME-2941: Make system_timer not use event memory
Save memory and reduce failure points by making single-shot
eventOS_event_timer_request allocate memory on the request call, and
pass this to the event system.

This means timers can no longer get lost at the time they are
scheduled to run. The only possibly failure is synchronous with the
request call.

However, the single-shot nature of the public API means we do have to
worry about failure on every request if trying to achieve periodic
timers - it's not possible to ensure we get a reliable repeated call.

As such not clear how much this will actually help in ONME-2941 in
isolation; more work will be required.