Fix post-event state change for user events
User-allocated events were marked as UNQUEUED after running. This is
would be incorrect if they were requeued while running, and that is
specifically permitted.

This would not normally cause a problem, but the state is used
in cancel(), so the following sequence would fail:

     Queue user-allocated event       QUEUED
      Event handler runs              RUNNING
       Event handler requeues event   QUEUED
      Event handler exits             UNQUEUED
     Someone cancels event

The event would not get cancelled due it being marked UNQUEUED.

Problem spotted by inspection while discussing using the "state"
variable in user code to check whether a user event was already
queued. This practice remains discouraged, as the arm_event_storage_t
is intended to be opaque, apart from its arm_event_s member.