Change event_data to uintptr_t
To aid people who want to pass a second pointer in an event, change
`event_data` from `uint32_t` to `uintptr_t`.

This is effectively a null change for 32-bit platforms, but will
add 8 bytes to the structure for 64-bit platforms (4 bytes of padding
and 4 bytes extra storage), and permit a pointer being stored there.

32-bit platforms might see some compiler warnings from `printf` or
similar if `uint32_t` is a different type to `uintptr_t` (`int` versus
`long`).

64-bit platforms will see more of a change, but expected problems are
minimal - on the whole it will just make 32-bit compatible
pointer-smuggling code that failed on 64-bit work.