Use event id 0 for initialization event
Some event handlers incorrectly check that event_id is 0 for initialization event. Even though
documentation does not mention that id is zeroed for initialization event, it is better to have some
init value instead of a random value. As eventloop has internal list of events, some old event id can
be used if id is not initialized.